+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Enhancements
 
-* Added support for using non-encrypted Realm in multiple processes from the same apk (#1091).
+* Added support for using non-encrypted Realms in multiple processes. Some caveats apply. Read [doc](https://realm.io/docs/java/latest/#multiprocess) for more info (#1091).
 
 ### Bug Fixes
