+++ examples/multiprocessExample/proguard-rules.pro
