+++ examples/multiprocessExample/src/main/AndroidManifest.xml
