+++ examples/multiprocessExample/src/main/res/values-w820dp/dimens.xml
