+++ examples/multiprocessExample/src/main/res/values/colors.xml
