+++ examples/multiprocessExample/src/main/res/values/dimens.xml
