+++ examples/multiprocessExample/src/main/res/values/strings.xml
