+++ examples/multiprocessExample/src/main/res/values/styles.xml
