+++ examples/settings.gradle
@@ -13,7 +13,7 @@ include 'unitTestExample'
 include 'newsreaderExample'
 include 'rxJavaExample'
 include 'objectServerExample'
-include 'multiProcessExample'
+include 'multiprocessExample'
 
 rootProject.name = 'realm-examples'
